## What's being changed

We've added a plugin to ensure B2B_Quotes in-progress imports are processed correctly. I've also fixed a warning that shows when running the dotdigital:sync NegotiableQuote CLI command. We've made this fix in our other modules previously.

## Why it's being changed

A refactor in the Email module has meant that this plugin is required. Currently, for merchants on newer versions (4.22.0+), B2B_Quotes import jobs will have had their data sent to Dotdigital, but will stay 'in-progress' forever and will not be checked and marked as either complete or failed.

## How to review / test this change

- Set up B2B quotes
- Create a quote
- Sync quotes
- Run the importer, twice
- All B2B_Quotes imports should be processed correctly and be either status 2 (complete) or 3 (failed)

Related work items: #270394

+++ Model/Sync/NegotiableQuote.php
@@ -135,7 +135,7 @@ class NegotiableQuote implements SyncInterface
 
         if ($this->countNegotiableQuotes) {
             $message = '----------- B2B Quotes sync ----------- : ' .
-                gmdate('H:i:s', microtime(true) - $this->start) .
+                gmdate('H:i:s', (int) microtime(true) - $this->start) .
                 ', Total synced = ' . $this->countNegotiableQuotes;
             $this->logger->info($message);
         }
